backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.24.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2